+++ skills-manifest.json
@@ -6,7 +6,7 @@
       "files": 144
     },
     "faceless-explainer": {
-      "hash": "4251c835adfda36b",
+      "hash": "ff47c02598dc7947",
       "files": 22
     },
     "figma": {
@@ -14,7 +14,7 @@
       "files": 2
     },
     "general-video": {
-      "hash": "c2efc32a529e682b",
+      "hash": "94fd399a0dd0b6ce",
       "files": 4
     },
     "hyperframes": {
@@ -30,8 +30,8 @@
       "files": 11
     },
     "hyperframes-core": {
-      "hash": "b30140bde17f4c7e",
-      "files": 18
+      "hash": "23f64febbc5cae37",
+      "files": 19
     },
     "hyperframes-creative": {
       "hash": "b9e2cbfa49e6ed6c",
@@ -46,7 +46,7 @@
       "files": 10
     },
     "media-use": {
-      "hash": "ef678a775bed5550",
+      "hash": "b10751149c9ea1da",
       "files": 145
     },
     "motion-graphics": {
@@ -58,11 +58,11 @@
       "files": 132
     },
     "pr-to-video": {
-      "hash": "7a0a69a86d8de15f",
+      "hash": "d37f9f5411c34089",
       "files": 29
     },
     "product-launch-video": {
-      "hash": "2e2dc6145d40b25a",
+      "hash": "10e0b80f7040ad1e",
       "files": 26
     },
     "remotion-to-hyperframes": {
